+++ tests/unzip_from_large_file.cc
@@ -17,12 +17,12 @@ TEST_F(PicounzipLargeFileTest, LargeFile) {
   // create_large_testdata.py
   {
     unzip zip("resource/large_data/large_entry.zip");
-    EXPECT_EQ(100000, zip.entrylist().size());
+    EXPECT_EQ(size_t(100000), zip.entrylist().size());
   }
   {
     unzip zip("resource/large_data/zero_largedata.zip");
-    EXPECT_EQ(uint64_t(1024) * 1024 * 1024 * 6,
-              zip.entrylist().front().file_size);
+    EXPECT_EQ(uint64_t(1024) * 1024 * 1024 * 6, uint64_t(
+              zip.entrylist().front().file_size));
 
 	unzip_file_stream unzifs(zip, zip.entrylist().front());
